+++ hois_back/src/test/java/ee/hitsa/ois/service/StateCurriculumCopyServiceTests.java
@@ -1,5 +1,7 @@
 package ee.hitsa.ois.service;
 
+import java.math.BigDecimal;
+
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -9,13 +11,13 @@ public class StateCurriculumCopyServiceTests {
 
     @Test
     public void calculateStudyPeriod() {
-        int result = StateCurriculumCopyService.calculateStudyPeriod(Long.valueOf(100));
+        int result = StateCurriculumCopyService.calculateStudyPeriod(BigDecimal.valueOf(100));
         Assert.assertTrue(20 == result);
     }
 
     @Test
     public void calculateStudyPeriodFraction() {
-        int result = StateCurriculumCopyService.calculateStudyPeriod(Long.valueOf(109));
+        int result = StateCurriculumCopyService.calculateStudyPeriod(BigDecimal.valueOf(109));
         Assert.assertTrue(22 == result);
     }
 }
